Add two different js gol.js and ising.js
The first is called in the Home Page and started the GoL simulation.
The second is called in the ising article. It is carry on a Metropolis
simulation of a Ising bidimensional lattice.

Was detected an issue on the pseudo-random numer generation performed with
Math.random(). Are there better ways to implement a PRNG in JavaScript?